+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Fixed
 
+- Dockerfile: `EXPOSE 12345` will now parse `12345` as an int instead of a string,
+  allowing `metavariable-comparison` with integers (#4875)
 - Scala: unicode character literals now parse
 - Scala: multiple annotated type parameters now parse (`def f[@an A, @an B](x : A, y : B) = ...`)